Rename prescription search response file
