+++ src/sandbox-spine-client.ts
@@ -2,7 +2,7 @@ import {AxiosResponse} from "axios"
 import {SpineClient} from "./spine-client"
 import {StatusCheckResponse} from "./status"
 import CLINICAL_CONTENT_VIEW_SANDBOX_RESPONSE from "./resources/clinical_content_view_sandbox"
-import PRESCRIPTION_SEARCH_RESPONSE from "./resources/prescription_search_response"
+import PRESCRIPTION_SEARCH_SANDBOX_RESPONSE from "./resources/prescription_search_sandbox"
 
 export class SandboxSpineClient implements SpineClient {
   async getStatus(): Promise<StatusCheckResponse> {
@@ -34,7 +34,7 @@ export class SandboxSpineClient implements SpineClient {
 
   async prescriptionSearch(): Promise<AxiosResponse> {
     const response: AxiosResponse = {
-      data: PRESCRIPTION_SEARCH_RESPONSE,
+      data: PRESCRIPTION_SEARCH_SANDBOX_RESPONSE,
       status: 200,
       statusText: "OK"
     } as unknown as AxiosResponse
